18th commit NDT
Thay đổi các lớp điều chỉnh 3 bảng sau: Product, Category, Image
Thêm 4 lỗi PRODUCT_NOT_FOUND, CATEGORY_EXIST, CATEGORY_NOT_FOUND, IMAGE_NOT_FOUND
Tắt JwtSecutiry
Sử dụng ModelMapper để map Product và User thay vì tạo package mapper riêng biệt

+++ back_end/src/main/java/com/example/back_end/repository/UserRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     boolean existsByUsername(String username);
     Optional<User> findUserByEmailAndPassword(String email, String password);
 
@@ -17,6 +17,6 @@ public interface UserRepository extends JpaRepository<User,Integer> {
 
     @Override
     <S extends User> List<S> findAll(Example<S> example);
-    Optional<User> findUserById(int id);
+    Optional<User> findUserById(Long id);
 }
